src/main/java/com/leetcode/tasks/easy/TwoSum.java
package com.leetcode.tasks.easy;

import java.util.HashMap;
import java.util.Map;

public class TwoSum {
    public int[] twoSum(int[] nums, int target) {
        return twoSumON(nums, target);
    }

    private int[] twoSumBrutForce(int[] nums, int target) {
        for (int i = 0; i < nums.length; i++) {
            for (int j = 0; j < nums.length; j++) {
                if (j == i) {
                    continue;
                }
                if (nums[j] + nums[i] == target) {
                    return new int[]{i, j};
                }
            }
        }

        return null;
    }

    private int[] twoSumON(int[] nums, int target) {
        Map<Integer, Integer> map = new HashMap<>();

        for (int i = 0; i < nums.length; i++) {
            Integer tmp = target - nums[i];
            if (map.containsKey(tmp)) {
                return new int[]{map.get(tmp), i};
            }
            map.put(nums[i], i);
        }

        return null;
    }
}



